Improved solution to challenge 1
The first solution generated a list of 4096 elements of all the
different base64 pairs, so that each group of 3 hexadecimal digits could
index that list directly by being interpreted as an integer.
This solution instead gets closer to reading the "raw bytes" of the hex
string by converting each pair of hex digits (= 1 byte) to an int, using
format() to convert that to an 8 digit binary representation, then
concatenating all of it.
Afterwards, the concatenated list of "bytes" is sliced so that we get
every 6 bits (base64 = 2^6) to directly index the base64 string.