Reorganising file naming and solution structures to ease future solutions
When trying to solve Challenge 3, I ran into some issues while
importing.

Issue 1: filenames started with a number. Since the `from x import y`
statements treated the filename of the module to be imported as a
regular identifier (i.e. not in quotes), having that filename start with
a digit made Python raise an error - any identifier that starts with a
digit is treated as a number. So I had to append "c" (for "challenge")
to all my solutions.

After the import worked, I realised another problem:

Issue 2: when a module is imported, everything inside it is executed,
even if it's outside the function you want to import.

This meant that any side effects (lines being printed, for example),
would occur before any logic in the script we were trying to run. Not
good. So I separated the purely functional modules with the actual
"solutions", where I check my answer.